+++ mRemoteNG/Connection/ConnectionInfoInheritance.cs
@@ -3,6 +3,7 @@ using System.ComponentModel;
 using System.Linq;
 using System.Reflection;
 using mRemoteNG.Tools;
+using mRemoteNG.Tree.Root;
 using mRemoteNG.Resources.Language;
 
 namespace mRemoteNG.Connection
@@ -455,7 +456,16 @@ namespace mRemoteNG.Connection
 		TypeConverter(typeof(MiscTools.YesNoTypeConverter))]public bool VNCViewOnly {get; set;}
         #endregion
 
-        [Browsable(false)] public ConnectionInfo Parent { get; private set; }
+        [Browsable(false)]
+        public ConnectionInfo Parent { get; private set; }
+
+        /// <summary>
+        /// Indicates whether this inheritance object is enabled.
+        /// When false, users of this object should not respect inheritance
+        /// settings for individual properties.
+        /// </summary>
+        [Browsable(false)]
+        public bool InheritanceActive => !(Parent is RootNodeInfo || Parent?.Parent is RootNodeInfo);
 
         #endregion
 
@@ -472,7 +482,6 @@ namespace mRemoteNG.Connection
         {
             var newInheritance = (ConnectionInfoInheritance)MemberwiseClone();
             newInheritance.Parent = parent;
-            newInheritance._tempInheritanceStorage = null;
             return newInheritance;
         }
 
@@ -530,15 +539,22 @@ namespace mRemoteNG.Connection
         /// <returns></returns>
         public IEnumerable<string> GetEnabledInheritanceProperties()
         {
-            return GetProperties()
-                .Where(property => (bool)property.GetValue(this))
-                .Select(property => property.Name)
-                .ToList();
+            return InheritanceActive
+                ? GetProperties()
+                    .Where(property => (bool)property.GetValue(this))
+                    .Select(property => property.Name)
+                    .ToList()
+                : Enumerable.Empty<string>();
         }
 
         private bool FilterProperty(PropertyInfo propertyInfo)
         {
-            var exclusions = new[] {"EverythingInherited", "Parent"};
+            var exclusions = new[]
+            {
+                nameof(EverythingInherited),
+                nameof(Parent),
+                nameof(InheritanceActive)
+            };
             var valueShouldNotBeFiltered = !exclusions.Contains(propertyInfo.Name);
             return valueShouldNotBeFiltered;
         }