+++ mRemoteNGTests/Connection/ConnectionInfoInheritanceTests.cs
@@ -1,77 +1,91 @@
-﻿using mRemoteNG.Connection;
-using mRemoteNG.Container;
-using NUnit.Framework;
-using System.Collections;
+﻿using System.Collections;
 using System.Linq;
 using System.Reflection;
+using mRemoteNG.Connection;
+using mRemoteNG.Container;
+using mRemoteNG.Tree.Root;
+using NUnit.Framework;
 
 namespace mRemoteNGTests.Connection
 {
-    [TestFixture]
+	[TestFixture]
     public class ConnectionInfoInheritanceTests
     {
-        private ConnectionInfo _connectionInfo;
-        private ConnectionInfoInheritance _inheritance;
-        private PropertyInfo[] _inheritanceProperties = typeof(ConnectionInfoInheritance).GetProperties();
+		private readonly PropertyInfo[] _inheritanceProperties = typeof(ConnectionInfoInheritance).GetProperties();
 
 
-        [SetUp]
-        public void Setup()
-        {
-            _connectionInfo = new ConnectionInfo();
-            _inheritance = new ConnectionInfoInheritance(_connectionInfo);
-        }
-
-        [TearDown]
-        public void Teardown()
-        {
-            _connectionInfo = null;
-            _inheritance = null;
-        }
-
-        [Test]
+		[Test]
         public void TurnOffInheritanceCompletely()
         {
-            _inheritance.Username = true;
-            _inheritance.TurnOffInheritanceCompletely();
-            Assert.That(AllInheritancePropertiesAreFalse(), Is.True);
+	        var inheritance = new ConnectionInfoInheritance(new ConnectionInfo()) {Username = true};
+	        inheritance.TurnOffInheritanceCompletely();
+            Assert.That(AllInheritancePropertiesAreFalse(inheritance), Is.True);
         }
 
         [Test]
         public void TurnOnInheritanceCompletely()
         {
-            _inheritance.TurnOnInheritanceCompletely();
-            Assert.That(AllInheritancePropertiesAreTrue(), Is.True);
+	        var inheritance = new ConnectionInfoInheritance(new ConnectionInfo());
+			inheritance.TurnOnInheritanceCompletely();
+            Assert.That(AllInheritancePropertiesAreTrue(inheritance), Is.True);
         }
 
         [Test]
-        public void DisableInheritanceTurnsOffAllInheritance()
+        public void InheritanceIsDisabledWhenAttachedToARootNode()
         {
-            _inheritance.Username = true;
-            _inheritance.DisableInheritance();
-            Assert.That(AllInheritancePropertiesAreFalse(), Is.True);
+	        var inheritance = new ConnectionInfoInheritance(new RootNodeInfo(RootNodeType.Connection));
+            Assert.That(inheritance.InheritanceActive, Is.False);
         }
 
         [Test]
-        public void EnableInheritanceRestoresPreviousInheritanceValues()
+        public void InheritanceIsDisabledWhenAttachedToAPuttyRootNode()
         {
-            _inheritance.Username = true;
-            _inheritance.DisableInheritance();
-            _inheritance.EnableInheritance();
-            Assert.That(_inheritance.Username, Is.True);
+	        var inheritance = new ConnectionInfoInheritance(new RootNodeInfo(RootNodeType.PuttySessions));
+	        Assert.That(inheritance.InheritanceActive, Is.False);
         }
 
         [Test]
+        public void InheritanceIsDisabledWhenAttachedToAPuttyNode()
+        {
+	        var inheritance = new ConnectionInfoInheritance(new RootPuttySessionsNodeInfo());
+	        Assert.That(inheritance.InheritanceActive, Is.False);
+        }
+
+		[Test]
+        public void InheritanceIsDisabledWhenAttachedToANodeDirectlyUnderTheRootNode()
+        {
+			var con = new ConnectionInfo();
+	        new RootNodeInfo(RootNodeType.Connection).AddChild(con);
+	        Assert.That(con.Inheritance.InheritanceActive, Is.False);
+        }
+
+		[Test]
+        public void InheritanceIsEnabledWhenAttachedToNormalConnectionInfo()
+        {
+	        var inheritance = new ConnectionInfoInheritance(new ConnectionInfo());
+            Assert.That(inheritance.InheritanceActive, Is.True);
+        }
+
+        [Test]
+        public void InheritanceIsEnabledWhenAttachedToNormalContainerInfo()
+        {
+	        var inheritance = new ConnectionInfoInheritance(new ContainerInfo());
+	        Assert.That(inheritance.InheritanceActive, Is.True);
+        }
+
+		[Test]
         public void GetPropertiesReturnsListOfSettableProperties()
         {
-            var hasIconProperty = _inheritance.GetProperties().Contains(typeof(ConnectionInfoInheritance).GetProperty("Icon"));
+	        var inheritance = new ConnectionInfoInheritance(new ConnectionInfo());
+			var hasIconProperty = inheritance.GetProperties().Contains(typeof(ConnectionInfoInheritance).GetProperty("Icon"));
             Assert.That(hasIconProperty, Is.True);
         }
 
         [Test]
         public void GetPropertiesExludesPropertiesThatShouldNotBeSet()
         {
-            var hasEverythingInheritedProperty = _inheritance.GetProperties().Contains(typeof(ConnectionInfoInheritance).GetProperty("EverythingInherited"));
+	        var inheritance = new ConnectionInfoInheritance(new ConnectionInfo());
+			var hasEverythingInheritedProperty = inheritance.GetProperties().Contains(typeof(ConnectionInfoInheritance).GetProperty("EverythingInherited"));
             Assert.That(hasEverythingInheritedProperty, Is.False);
         }
 
@@ -91,23 +105,25 @@ namespace mRemoteNGTests.Connection
             Assert.That(con1.AutomaticResize, Is.EqualTo(expectedSetting));
         }
 
-        private bool AllInheritancePropertiesAreTrue()
+        private bool AllInheritancePropertiesAreTrue(ConnectionInfoInheritance inheritance)
         {
             var allPropertiesTrue = true;
             foreach (var property in _inheritanceProperties)
             {
-                if (PropertyIsBoolean(property) && PropertyIsChangedWhenSettingInheritAll(property) && BooleanPropertyIsSetToFalse(property))
+                if (PropertyIsBoolean(property) && PropertyIsChangedWhenSettingInheritAll(property) && 
+                    BooleanPropertyIsSetToFalse(property, inheritance))
                     allPropertiesTrue = false;
             }
             return allPropertiesTrue;
         }
 
-        private bool AllInheritancePropertiesAreFalse()
+        private bool AllInheritancePropertiesAreFalse(ConnectionInfoInheritance inheritance)
         {
             var allPropertiesFalse = true;
             foreach (var property in _inheritanceProperties)
             {
-                if (PropertyIsBoolean(property) && PropertyIsChangedWhenSettingInheritAll(property) && BooleanPropertyIsSetToTrue(property))
+                if (PropertyIsBoolean(property) && PropertyIsChangedWhenSettingInheritAll(property) && 
+                    BooleanPropertyIsSetToTrue(property, inheritance))
                     allPropertiesFalse = false;
             }
             return allPropertiesFalse;
@@ -124,14 +140,14 @@ namespace mRemoteNGTests.Connection
             return (property.PropertyType.Name == typeof(bool).Name);
         }
 
-        private bool BooleanPropertyIsSetToFalse(PropertyInfo property)
+        private bool BooleanPropertyIsSetToFalse(PropertyInfo property, ConnectionInfoInheritance inheritance)
         {
-            return (bool)property.GetValue(_inheritance) == false;
+            return (bool)property.GetValue(inheritance) == false;
         }
 
-        private bool BooleanPropertyIsSetToTrue(PropertyInfo property)
+        private bool BooleanPropertyIsSetToTrue(PropertyInfo property, ConnectionInfoInheritance inheritance)
         {
-            return (bool)property.GetValue(_inheritance);
+            return (bool)property.GetValue(inheritance);
         }
     }
 }